LayoutTests/editing/execCommand/indent-nested-lists-1-expected.txt

EDITING DELEGATE: shouldBeginEditingInDOMRange:range from 0 of OL > DIV > BODY > HTML > #document to 9 of OL > DIV > BODY > HTML > #document
EDITING DELEGATE: webViewDidBeginEditing:WebViewDidBeginEditingNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: shouldChangeSelectedDOMRange:range from 6 of OL > DIV > BODY > HTML > #document to 7 of OL > DIV > BODY > HTML > #document toDOMRange:range from 0 of #text > LI > OL > OL > DIV > BODY > HTML > #document to 5 of #text > LI > OL > OL > DIV > BODY > HTML > #document affinity:NSSelectionAffinityDownstream stillSelecting:FALSE
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChange:WebViewDidChangeNotification
This tests indenting "three". You should see numbers 1 and 2 before "three" and "four" respectively.
one
two
three
four
Before: <li>one</li> <li>two</li> <li id="test">three</li> <ol><li>four</li></ol>
After: <li>one</li> <li>two</li> <ol><li id="test">three</li><li>four</li></ol>
